Orda_Flutter/simple_jeju_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
langchain-upstage==0.1.8
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0 